#148 Add missing events and fix scaling
